requirements.txt
absl-py==0.9.0
allennlp==1.1.0rc1.dev20200709
allennlp-models==1.1.0rc1.dev20200709
astroid==2.4.2
attrs==19.3.0
blis==0.4.1
boto3==1.14.20
botocore==1.17.20
cachetools==4.0.0
catalogue==1.0.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
conllu==3.0
cycler==0.10.0
cymem==2.0.3
docutils==0.15.2
easydict==1.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
filelock==3.0.12
Flask==1.1.2
future==0.18.2
google-auth==1.13.1
google-auth-oauthlib==0.4.1
grpcio==1.28.1
gym==0.17.1
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.7.0
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonnet==0.16.0
jsonpickle==1.4.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
Metaphone==0.6
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
more-itertools==8.4.0
murmurhash==1.0.2
nltk==3.4.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.2.0.34
overrides==3.1.0
packaging==20.4
Pillow @ file:///tmp/build/80754af9/pillow_1603822253009/work
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
pronounceable==0.1.3
protobuf==3.11.3
py==1.9.0
py-rouge==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pylint==2.6.0
pyparsing==2.4.6
PyQt5==5.14.2
PyQt5-sip==12.7.2
pytest==5.4.3
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
sentencepiece==0.1.91
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sniffio==1.2.0
spacy==2.2.4
srsly==1.0.2
stanza==1.1.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.1
thinc==7.4.0
threadpoolctl==2.1.0
tokenizers==0.7.0
toml==0.10.1
torch==1.7.1
torchvision==0.8.2
tqdm==4.45.0
transformers==2.11.0
typed-ast==1.4.1
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
urllib3==1.25.8
wasabi==0.7.0
wcwidth==0.2.5
Werkzeug==1.0.1
word2number==1.1
wrapt==1.12.1
xlwt==1.3.0
zipp==3.1.0
